Complete the algorithm and add regularization

+++ LinearRegression/CostFunction.cpp
@@ -5,7 +5,7 @@ double CostFunction::MSE(vector<double> yPred, vector<double> yTrue)
     double cost = 0.0;
     int m = yTrue.size();
 
-    for (int i = 0; i < yTrue.size(); i++)
+    for (size_t i = 0; i < yTrue.size(); i++)
     {
         double diff = yPred[i] - yTrue[i];
         cost += diff * diff;
